.history/requirements_20250409123130.txt
langchain==0.3.9
langchain-neo4j==0.4.0
langchain-google-genai==0.0.12
neo4j==5.27.0
google-generativeai==0.3.2
python-dotenv==1.0.1
